+++ servers/mcp-neo4j-data-modeling/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added
 * Add HTTP transport option
+* Migrate to FastMCP v2.x
 
 ## v0.1.1